requirements.txt
aiogram==2.4
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
heroku==0.1.4
idna==2.8
idna-ssl==1.1.0
git+https://github.com/alberanid/imdbpy
kinopoiskpy==0.8.2
lxml==4.4.2
multidict==4.7.2
python-dateutil==1.5
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
typing-extensions==3.7.4.1
urllib3==1.25.7
yarl==1.4.2
